+++ src/components/App/app.jsx
@@ -1,0 +1,7 @@
+
+ function App() {
+        return <h1>Hello World!</h1>;
+      }
+export default App;
+      
+
